// challenges/github-user-search/src/types/UserData.types.ts
export type UserData = {
    name: string | null;
    login: string;
    createdAt: string;
    avatarUrl: string;
    location: string | null;
    twitterUsername: string | null;
    company: string | null;
    blog: string | null;
    bio: string | null;
    publicRepos: number;
    followers: number;
    following: number;
};